test: move check function tests to new file
